e2e: port OTP CAPI tests to OTE framework and skip on External topology
Move tests from operator.go and webhook_validation.go into
feature-gate-aligned files: cluster_api_machine_management.go for
common tests, and cluster_api_machine_management_{aws,gcp,vsphere}.go
for platform-specific secret sync tests.

Add Label("skip-topology:External") and BeforeEach skip guards to the
ported OTP Describe blocks whose originals carried NonHyperShiftHOST
in openshift-tests-private (OTP IDs: 61980, 71913, 73620, 76078,
55205). Secret sync tests (OTP 51141) are left unchanged as the
original had no NonHyperShiftHOST marker.
